Added more links to edit the boost
I did not think there was a way to edit boosts because there was no link on the
page to access it.  Only after tracing the code did I notice the index column
for the admin panel was actually linked to the edit function.  I added the
link to the name and strength columns as well so that the edit function was
easier to find.

(cherry picked from commit 99f33fd8e336fb31dafa2ab948c57dec88964a71)